refactored the code by adding sensor class and added MoistureSensor class
